requirement.txt
beautifulsoup4==4.6.3
certifi==2018.8.13
chardet==3.0.4
Django==2.1
django-request-logging==0.6.5
djangorestframework==3.8.2
drf-tracking==1.4.0
idna==2.7
pytz==2018.5
requests==2.19.1
urllib3==1.23
